inline the sample code
